backend/pyproject.toml
[tool.poetry]
name = "bibliogon"
version = "0.19.0"
description = "Open-source book authoring platform with WYSIWYG and Markdown editing, EPUB/PDF export via Pandoc."
authors = ["Asterios Raptis"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.135.0"
uvicorn = { extras = ["standard"], version = "^0.44.0" }
sqlalchemy = "^2.0.49"
pydantic = "^2.13.0"
python-multipart = "^0.0.26"
pluginforge = {version = "^0.5.0", extras = ["fastapi"]}
bibliogon-plugin-export = {path = "../plugins/bibliogon-plugin-export", develop = true}
bibliogon-plugin-help = {path = "../plugins/bibliogon-plugin-help", develop = true}
bibliogon-plugin-getstarted = {path = "../plugins/bibliogon-plugin-getstarted", develop = true}
bibliogon-plugin-ms-tools = {path = "../plugins/bibliogon-plugin-ms-tools", develop = true}
bibliogon-plugin-audiobook = {path = "../plugins/bibliogon-plugin-audiobook", develop = true}
bibliogon-plugin-grammar = {path = "../plugins/bibliogon-plugin-grammar", develop = true}
bibliogon-plugin-kdp = {path = "../plugins/bibliogon-plugin-kdp", develop = true}
bibliogon-plugin-kinderbuch = {path = "../plugins/bibliogon-plugin-kinderbuch", develop = true}
bibliogon-plugin-translation = {path = "../plugins/bibliogon-plugin-translation", develop = true}
markdown = "^3.10.2"
alembic = "^1.18.4"
pillow = "^12.0.0"
gtts = "^2.5.0"
pyttsx3 = "^2.90"
elevenlabs = "^0.2.27"
cryptography = "^46.0.7"
ruamel-yaml = "^0.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
httpx = "^0.28.0"
mutmut = "^3.5.0"
mypy = "^1.20.0"
types-pyyaml = "^6.0.12.20250915"
types-markdown = "^3.10.2.20260211"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.1.0"
ruff = "^0.15.11"
pre-commit = "^4.5.1"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Conservative baseline. Tighten in follow-ups once the code is clean
# against the current set.
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",   # line length - handled by formatter, too noisy to fix blind
    "B008",   # function-call-in-default-argument - FastAPI `Depends(...)` idiom is intentional
    "UP042",  # replace-str-enum - `(str, Enum)` base class ships in migrations too; deferred
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B"]                       # bugbear too strict for test fixtures
"migrations/versions/*" = ["UP"]        # Alembic templates predate py312 syntax
"app/main.py" = ["E402"]                # setup_logging() before imports is intentional
"app/licensing.py" = ["E402"]           # LICENSING_ENABLED flag before imports is intentional

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = "pluginforge.*"
ignore_missing_imports = true

# bibliogon_audiobook + manuscripta are the only plugin modules
# actually imported from backend/app/ at runtime (audiobook router
# delegates into them). Other plugins (export, grammar, kdp, ms_tools,
# kinderbuch, translation, help, getstarted) are loaded dynamically
# via the plugin manager, never imported from `app/`, so their
# mypy overrides would be unused.
[[tool.mypy.overrides]]
module = "bibliogon_audiobook.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "manuscripta.*"
ignore_missing_imports = true

[tool.mutmut]
paths_to_mutate = ["app/"]
tests_dir = ["tests/"]
do_not_mutate = ["app/main.py", "app/database.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
